refactor(appointment-details): refactor code for dialogs, remove unnecessary code
