+++ src/utils/dateUtils.js
@@ -1,4 +1,4 @@
-export const formatBirthdate = (dateString) => {
+export const formatDate = (dateString) => {
   const date = new Date(dateString);
   return date.toLocaleDateString("en-US", {
     month: "long",
@@ -7,12 +7,11 @@ export const formatBirthdate = (dateString) => {
   });
 };
 
-export const formatDate = (dateString) => {
-  const date = new Date(dateString);
-  return date.toLocaleDateString("en-US", {
-    month: "long",
-    day: "numeric",
-    year: "numeric",
+export const formatTime = (timeString) => {
+  return new Date(timeString).toLocaleTimeString("en-US", {
+    hour: "2-digit",
+    minute: "2-digit",
+    hour12: true,
   });
 };
